src/backend/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
pydantic-settings
pydantic[email]
python-dotenv
psycopg2-binary
numpy
phonenumbers
pgvector
pytest
